Globally refactor button debounce

+++ src/lights/led-strip.cpp
@@ -19,7 +19,7 @@ namespace Lights
 
   void LedStrip::reset()
   {
-    for (int i = 0; i < size(); i++)
+    for (uint16_t i = 0; i < size(); i++)
     {
       buffer[i] = 0;
     }
@@ -28,7 +28,7 @@ namespace Lights
   void LedStrip::adjustLuminance()
   {
     luminance.adjustLuminance();
-    for (int i = 0; i < size(); ++i)
+    for (uint16_t i = 0; i < size(); ++i)
     {
       float scale = static_cast<float>(luminance.getLuminance()) / LedLuminance::MAX_LED_BRIGHTNESS;
       buffer[i] = buffer[i] * scale;
